refactor(utils): extract internal metadata utils to public API
