+++ src/AddressDriver/AddressDriverPresets.ts
@@ -9,7 +9,7 @@ import {
 	validateSplitInput,
 	validateSqueezeDripsInput
 } from '../common/validators';
-import { createFromStrings, isNullOrUndefined, nameOf } from '../common/internals';
+import { isNullOrUndefined, nameOf } from '../common/internals';
 import Utils from '../utils';
 import type { DripsReceiverStruct, Preset, SplitsReceiverStruct, SqueezeArgs, UserMetadata } from '../common/types';
 import { DripsErrors } from '../common/DripsError';
@@ -131,7 +131,7 @@ export namespace AddressDriverPresets {
 				transferToAddress
 			);
 
-			const userMetadataAsBytes = userMetadata.map((m) => createFromStrings(m.key, m.value));
+			const userMetadataAsBytes = userMetadata.map((m) => Utils.Metadata.createFromStrings(m.key, m.value));
 
 			const emitUserMetadataTx = await addressDriverTxFactory.emitUserMetadata(userMetadataAsBytes);
