+++ src/common/internals.ts
@@ -3,10 +3,10 @@
 /* eslint-disable no-restricted-syntax */
 /* eslint-disable no-nested-ternary */
 
-import type { BytesLike, PopulatedTransaction, Signer } from 'ethers';
-import { BigNumber, ethers } from 'ethers';
+import type { PopulatedTransaction, Signer } from 'ethers';
+import { BigNumber } from 'ethers';
 import { DripsErrors } from './DripsError';
-import type { DripsReceiverStruct, SplitsReceiverStruct, UserMetadataStruct } from './types';
+import type { DripsReceiverStruct, SplitsReceiverStruct } from './types';
 
 /** @internal */
 export const nameOf = (obj: any) => Object.keys(obj)[0];
@@ -75,38 +75,6 @@ export function ensureSignerExists(signer: Signer | undefined): asserts signer i
 		throw DripsErrors.signerMissingError();
 	}
 }
-
-/** @internal */
-export const keyFromString = (key: string): BytesLike => ethers.utils.formatBytes32String(key);
-
-/** @internal */
-export const valueFromString = (value: string): BytesLike => ethers.utils.hexlify(ethers.utils.toUtf8Bytes(value));
-
-/** @internal */
-export const createFromStrings = (
-	key: string,
-	value: string
-): {
-	key: BytesLike;
-	value: BytesLike;
-} => ({
-	key: keyFromString(key),
-	value: valueFromString(value)
-});
-
-/** @internal */
-export const parseMetadataAsString = (userMetadata: UserMetadataStruct): { key: string; value: string } => {
-	if (!ethers.utils.isBytesLike(userMetadata?.key) || !ethers.utils.isBytesLike(userMetadata?.value)) {
-		throw DripsErrors.argumentError(
-			`Invalid key-value user metadata pair: key or value is not a valid BytesLike object.`
-		);
-	}
-
-	return {
-		key: ethers.utils.parseBytes32String(userMetadata.key),
-		value: ethers.utils.toUtf8String(userMetadata.value)
-	};
-};
 
 /**
  * Source: https://github.com/radicle-dev/drips-app-v2/blob/main/src/lib/utils/expect.ts
